+++ org/org.go
@@ -24,7 +24,6 @@ func init() {
 		Person{},
 		Registration{},
 		Telephone{},
-		Unit(""),
 		Website{},
 		Attachment{},
 	)
@@ -46,7 +45,6 @@ func init() {
 		personRules(),
 		registrationRules(),
 		telephoneRules(),
-		unitRules(),
 		websiteRules(),
 	)
 }